refactor and cleaanup
